+++ metadata.rb
@@ -8,11 +8,12 @@ long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.4.0'
 # rubocop:enable Style/SingleSpaceBeforeFirstArg
 
-%w( ubuntu ).each do |os|
+%w( ubuntu debian centos redhat fedora ).each do |os|
   supports os
 end
 
 depends 'apt'
+depends 'yum'
 depends 'build-essential'
 depends 'runit', '~> 1.5'
 depends 'bluepill', '~> 2.3'