+++ recipes/source.rb
@@ -19,33 +19,19 @@
 
 include_recipe 'build-essential::default'
 
-user node['prometheus']['user'] do
-  system    true
-  shell     '/bin/false'
-  home      node['prometheus']['dir']
-  not_if    { node['prometheus']['use_existing_user'] == true || node['prometheus']['user'] == "root" }
-end
-
-directory node['prometheus']['dir'] do
-  owner     node['prometheus']['user']
-  group     node['prometheus']['group']
-  mode      '0755'
-  recursive true
-end
-
 # These packages are needed go build
 %w( curl git-core mercurial gzip sed ).each do |pkg|
   package pkg
 end
 
-git "#{Chef::Config[:file_cache_path]}/prometheus-#{node['prometheus']['source']['version']}" do
+git "#{Chef::Config[:file_cache_path]}/prometheus-#{node['prometheus']['version']}" do
   repository node['prometheus']['source']['git_repository']
   revision node['prometheus']['source']['git_revision']
   action :checkout
 end
 
 bash 'compile_prometheus_source' do
-  cwd "#{Chef::Config[:file_cache_path]}/prometheus-#{node['prometheus']['source']['version']}"
+  cwd "#{Chef::Config[:file_cache_path]}/prometheus-#{node['prometheus']['version']}"
   code <<-EOH
     make build &&
     cp -R prometheus #{node['prometheus']['dir']} &&
@@ -59,55 +45,3 @@ bash 'compile_prometheus_source' do
 
   notifies :restart, 'service[prometheus]'
 end
-
-# this will only get written here the first time
-template node['prometheus']['flags']['config.file'] do
-  action :create_if_missing
-  cookbook node['prometheus']['job_config_cookbook_name']
-  source node['prometheus']['job_config_template_name']
-  mode 0644
-  owner node['prometheus']['user']
-  group node['prometheus']['group']
-  notifies :restart, 'service[prometheus]'
-end
-
-directory node['prometheus']['log_dir'] do
-  owner node['prometheus']['user']
-  group node['prometheus']['group']
-  mode '0755'
-  recursive true
-end
-
-case node['prometheus']['init_style']
-when 'runit'
-  include_recipe 'runit::default'
-
-  runit_service 'prometheus' do
-    default_logger true
-  end
-when 'bluepill'
-  include_recipe 'bluepill::default'
-
-  template "#{node['bluepill']['conf_dir']}/prometheus.pill" do
-    source 'prometheus.pill.erb'
-    mode 0644
-  end
-
-  bluepill_service 'prometheus' do
-    action [:enable, :load]
-  end
-else
-  template '/etc/init.d/prometheus' do
-    source 'prometheus.erb'
-    owner 'root'
-    group node['root_group']
-    mode '0755'
-    notifies :restart, 'service[prometheus]', :delayed
-  end
-end
-
-# rubocop:disable Style/HashSyntax
-service 'prometheus' do
-  supports :status => true, :restart => true
-end
-# rubocop:enable Style/HashSyntax